Geek-Academy-Spec-Driven-workshop-main/support-agent-python/requirements.txt
python-dotenv==1.2.2
pytest==7.4.3
pytest-asyncio==0.21.1
azure-ai-inference==1.0.0b2
azure-identity==1.14.0
